requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.5
django-simple-email-confirmation==0.70
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
